requirements.txt
pymysql==1.0.2
pandas==1.4.3
psycopg2_binary==2.9.3
SQLAlchemy==1.4.39
